debris is if now

+++ code/modules/cm_marines/dropship_ammo.dm
@@ -318,27 +318,22 @@
 	fire_mission_delay = 0 //0 means unusable
 
 /obj/structure/ship_ammo/rocket/napalm/detonate_on(turf/impact)
-	  	impact.ceiling_debris_check(3)
+	  	if(impact && CEILING_IS_PROTECTED(A?.ceiling, CEILING_PROTECTION_TIER_3) && protected_by_pylon(TURF_PROTECTION_CAS, impact)
 			var/list/turf_list = list()
 			for(var/turf/T in range(impact, 7)) //This is its area of effect
 			turf_list += T
 			spawn(5)
 				cell_explosion(impact, 100, 25, EXPLOSION_FALLOFF_SHAPE_LINEAR, null, create_cause_data(initial(name), source_mob))
-		  		for(var/i=1 to 26) //This is how many tiles within that area of effect will be randomly ignited on instant
-					var/turf/U = pick(turf_list)
-					turf_list -= U
-					spawn(5)
-						fire_spread(U, create_cause_data(initial(name), source_mob), 4, 30, 25, "#EE6515")
-				qdel(src)
-		impact.ceiling_debris_check(2)
+		  	for(var/i=1 to 26) //This is how many tiles within that area of effect will be randomly ignited on instant
+				var/turf/U = pick(turf_list)
+				turf_list -= U
+				spawn(5)
+					fire_spread(U, create_cause_data(initial(name), source_mob), 4, 30, 25, "#EE6515")
+			qdel(src)
+		else
 			spawn(5)
 				cell_explosion(impact, 150, 25, EXPLOSION_FALLOFF_SHAPE_LINEAR, null, create_cause_data(initial(name), source_mob))
-				fire_spread(impact, create_cause_data(initial(name), source_mob), 4, 30, 30, "#EE6515")
-			qdel(src)
-		impact.ceiling_debris_check(1)
-			spawn(5)
-				cell_explosion(impact, 200, 25, EXPLOSION_FALLOFF_SHAPE_LINEAR, null, create_cause_data(initial(name), source_mob))
-				fire_spread(impact, create_cause_data(initial(name), source_mob), 6, 60, 30, "#EE6515")
+				fire_spread(impact, create_cause_data(initial(name), source_mob), 4, 60, 40, "#EE6515")
 		qdel(src)
 
 
@@ -423,8 +418,8 @@
 		var/datum/effect_system/smoke_spread/S = new/datum/effect_system/smoke_spread()
 		S.set_up(5,0,impact,null,30)
 		S.start
-		if(!ammo_count && loc)
-			qdel(src)
+	if(!ammo_count && loc)
+		qdel(src)
 
 
             qdel(src)
